:recycle:refactor(controller): Controller 싱글톤, 정적클래스 설정 변경
- FirstController: run() 메소드가 딱 한 번, 무조건 실행되어야 함.
- 따라서 싱글톤으로 만든 뒤 객체 생성 시에만 해당 메소드 실행되도록 함.
- ControllerHandler: FirstController의 내부 이넘 클래스로 변경하여 원치 않는 실행 방지.
- 세부 기능 Controller: 공개 메소드의 접근제어자 default로 변경. (controller 패키지에서만 접근 가능)
- 독립적으로 각 도메인 별 기능을 관리하면서 정적 메소드들을 객체처럼 쓰도록 함.
