+++ src/main/java/subway/Application.java
@@ -4,6 +4,6 @@ import subway.controller.FirstController;
 
 public class Application {
     public static void main(String[] args) {
-        FirstController.run();
+        FirstController.initializeInstance();
     }
 }
