+++ src/main/java/subway/controller/ErrorInterceptor.java
@@ -4,6 +4,9 @@ import subway.view.OutputView;
 
 public class ErrorInterceptor {
 
+    private ErrorInterceptor() {
+    }
+
     public static void runUntilGetLegalArguments(Runnable menuService) {
         while (true) {
             try {
