+++ src/main/java/subway/controller/FirstController.java
@@ -4,35 +4,73 @@ import static subway.controller.RunStatus.RUNNING;
 import static subway.controller.RunStatus.STOPPED;
 import static subway.view.constants.menu.MainCommand.QUIT;
 
+import java.util.Arrays;
+import java.util.Map;
+import java.util.stream.Collectors;
 import subway.view.InputView;
 import subway.view.OutputView;
 import subway.view.constants.menu.MainCommand;
 
 public class FirstController {
 
-    private static RunStatus runStatus = STOPPED;
+    private static FirstController instance;
+    private RunStatus runStatus = STOPPED;
 
     private FirstController() {
+        run();
     }
 
-    public static void run() {
-        runStatus = RUNNING;
-        while (runStatus == RUNNING) {
-            ErrorInterceptor.runUntilGetLegalArguments(FirstController::selectMainMenu);
+    public static void initializeInstance() {
+        if (instance == null) {
+            instance = new FirstController();
         }
     }
 
-    private static void selectMainMenu() {
+    private void run() {
+        runStatus = RUNNING;
+        while (runStatus == RUNNING) {
+            ErrorInterceptor.runUntilGetLegalArguments(this::selectMainMenu);
+        }
+    }
+
+    private void selectMainMenu() {
         OutputView.printMainMenus();
         MainCommand mainCommand = InputView.inputMainCommand();
         runSelectedController(mainCommand);
     }
 
-    private static void runSelectedController(MainCommand mainCommand) {
+    private void runSelectedController(MainCommand mainCommand) {
         if (mainCommand == QUIT) {
             runStatus = STOPPED;
             return;
         }
         ControllerHandler.run(mainCommand);
+    }
+
+    private enum ControllerHandler {
+
+        STATION(MainCommand.STATION, StationController::selectMenu),
+        LINE(MainCommand.LINE, LineController::selectMenu),
+        PATH(MainCommand.PATH, PathController::selectMenu),
+        SUBWAY_LINES(MainCommand.SUBWAY_LINES, PathController::showAllSubwayLines),
+        ;
+
+        private static final Map<MainCommand, Runnable> runnableByMainCommand = Arrays.stream(values())
+                .collect(Collectors.toMap(value -> value.mainCommand, value -> value.runnable));
+        private final MainCommand mainCommand;
+        private final Runnable runnable;
+
+        ControllerHandler(MainCommand mainCommand, Runnable runnable) {
+            this.mainCommand = mainCommand;
+            this.runnable = runnable;
+        }
+
+        public static void run(MainCommand command) {
+            Runnable controllerMethod = runnableByMainCommand.get(command);
+            if (controllerMethod == null) {
+                throw new IllegalArgumentException("해당 입력 키워드로 실행할 수 있는 기능이 없습니다.");
+            }
+            ErrorInterceptor.runUntilGetLegalArguments(controllerMethod);
+        }
     }
 }
